Moved code to solve dependencies of salesforce.backend
File salesforce/backend/__init__.py can be imported now before
django.setup() is configured.
