+++ salesforce/auth.py
@@ -22,7 +22,7 @@ import logging
 import requests
 import threading
 from django.db import connections
-from salesforce.backend import MAX_RETRIES
+from salesforce.backend import get_max_retries
 from salesforce.backend.driver import DatabaseError, IntegrityError
 from salesforce.backend.adapter import SslHttpAdapter
 from requests.auth import AuthBase
@@ -173,7 +173,7 @@ class SalesforcePasswordAuth(SalesforceAuth):
         url = ''.join([settings_dict['HOST'], '/services/oauth2/token'])
 
         log.info("attempting authentication to %s" % settings_dict['HOST'])
-        self._session.mount(settings_dict['HOST'], SslHttpAdapter(max_retries=MAX_RETRIES))
+        self._session.mount(settings_dict['HOST'], SslHttpAdapter(max_retries=get_max_retries()))
         response = self._session.post(url, data=dict(
             grant_type      = 'password',
             client_id       = settings_dict['CONSUMER_KEY'],
